+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added enum variant `State::Open`
 - Implemented `std::std::FromStr` for all task components
 - Added rustdoc documentation to the public api
-- Added feature `chrono` for optional support for `chrono::Local` dates
+- Added feature `chrono` for optional support for `chrono::NaiveDate` dates
 - Added feature `serde` for optional support for de(serializing) todo.txt components
 
 ### Changed